Update all other packages to the new conventions
