--- packages/workbox-broadcast-cache-update/_public.mjs
@@ -1,20 +1,0 @@
-/*
-  Copyright 2018 Google LLC
-
-  Use of this source code is governed by an MIT-style
-  license that can be found in the LICENSE file or at
-  https://opensource.org/licenses/MIT.
-*/
-
-import {BroadcastCacheUpdate} from './BroadcastCacheUpdate.mjs';
-import {Plugin} from './Plugin.mjs';
-import {broadcastUpdate} from './broadcastUpdate.mjs';
-import {responsesAreSame} from './responsesAreSame.mjs';
-import './_version.mjs';
-
-export {
-  BroadcastCacheUpdate,
-  Plugin,
-  broadcastUpdate,
-  responsesAreSame,
-};
