--- packages/workbox-cacheable-response/_public.mjs
@@ -1,16 +1,0 @@
-/*
-  Copyright 2018 Google LLC
-
-  Use of this source code is governed by an MIT-style
-  license that can be found in the LICENSE file or at
-  https://opensource.org/licenses/MIT.
-*/
-
-import {CacheableResponse} from './CacheableResponse.mjs';
-import {Plugin} from './Plugin.mjs';
-import './_version.mjs';
-
-export {
-  CacheableResponse,
-  Plugin,
-};
