+++ packages/workbox-core/index.mjs
@@ -8,7 +8,7 @@
 
 import {registerQuotaErrorCallback} from './registerQuotaErrorCallback.mjs';
 import * as _private from './_private.mjs';
-import {clientClaim} from './clientClaim.mjs';
+import {clientsClaim} from './clientsClaim.mjs';
 import {getCacheNames} from './getCacheNames.mjs';
 import {setCacheNameDetails} from './setCacheNameDetails.mjs';
 import {skipWaiting} from './skipWaiting.mjs';
@@ -33,7 +33,7 @@ try {
 
 export {
   _private,
-  clientClaim,
+  clientsClaim,
   getCacheNames,
   registerQuotaErrorCallback,
   setCacheNameDetails,
