--- packages/workbox-navigation-preload/_public.mjs
@@ -1,19 +1,0 @@
-/*
-  Copyright 2018 Google LLC
-
-  Use of this source code is governed by an MIT-style
-  license that can be found in the LICENSE file or at
-  https://opensource.org/licenses/MIT.
-*/
-
-import {disable} from './disable.mjs';
-import {enable} from './enable.mjs';
-import {isSupported} from './utils/isSupported.mjs';
-
-import './_version.mjs';
-
-export {
-  disable,
-  enable,
-  isSupported,
-};
