--- packages/workbox-precaching/_public.mjs
@@ -1,15 +1,0 @@
-/*
-  Copyright 2018 Google LLC
-
-  Use of this source code is governed by an MIT-style
-  license that can be found in the LICENSE file or at
-  https://opensource.org/licenses/MIT.
-*/
-
-import PrecacheController from './controllers/PrecacheController.mjs';
-
-import './_version.mjs';
-
-export {
-  PrecacheController,
-};
