+++ packages/workbox-precaching/utils/getOrCreatePrecacheController.mjs
@@ -1,0 +1,23 @@
+/*
+  Copyright 2019 Google LLC
+
+  Use of this source code is governed by an MIT-style
+  license that can be found in the LICENSE file or at
+  https://opensource.org/licenses/MIT.
+*/
+
+import {PrecacheController} from '../PrecacheController.mjs';
+import '../_version.mjs';
+
+
+let precacheController;
+
+/**
+ * @return {PrecacheController}
+ */
+export const getOrCreatePrecacheController = () => {
+  if (!precacheController) {
+    precacheController = new PrecacheController();
+  }
+  return precacheController;
+};
