+++ packages/workbox-precaching/utils/precachePlugins.mjs
@@ -1,0 +1,30 @@
+/*
+  Copyright 2019 Google LLC
+
+  Use of this source code is governed by an MIT-style
+  license that can be found in the LICENSE file or at
+  https://opensource.org/licenses/MIT.
+*/
+
+import '../_version.mjs';
+
+
+let plugins = [];
+
+export const precachePlugins = {
+  /*
+   * @return {Array}
+   * @private
+   */
+  get() {
+    return [...plugins];
+  },
+
+  /*
+   * @param {Array} newPlugins
+   * @private
+   */
+  add(newPlugins) {
+    plugins.push(...newPlugins);
+  },
+};
