+++ packages/workbox-strategies/CacheFirst.mjs
@@ -14,7 +14,7 @@ import {getFriendlyURL} from 'workbox-core/_private/getFriendlyURL.mjs';
 import {logger} from 'workbox-core/_private/logger.mjs';
 import {WorkboxError} from 'workbox-core/_private/WorkboxError.mjs';
 
-import messages from './utils/messages.mjs';
+import {messages} from './utils/messages.mjs';
 import './_version.mjs';
 
 /**
