+++ packages/workbox-strategies/NetworkOnly.mjs
@@ -12,7 +12,7 @@ import {fetchWrapper} from 'workbox-core/_private/fetchWrapper.mjs';
 import {logger} from 'workbox-core/_private/logger.mjs';
 import {WorkboxError} from 'workbox-core/_private/WorkboxError.mjs';
 
-import messages from './utils/messages.mjs';
+import {messages} from './utils/messages.mjs';
 import './_version.mjs';
 
 /**
