--- packages/workbox-streams/_public.mjs
@@ -1,21 +1,0 @@
-/*
-  Copyright 2018 Google LLC
-
-  Use of this source code is governed by an MIT-style
-  license that can be found in the LICENSE file or at
-  https://opensource.org/licenses/MIT.
-*/
-
-import {concatenate} from './concatenate.mjs';
-import {concatenateToResponse} from './concatenateToResponse.mjs';
-import {isSupported} from './isSupported.mjs';
-import {strategy} from './strategy.mjs';
-
-import './_version.mjs';
-
-export {
-  concatenate,
-  concatenateToResponse,
-  isSupported,
-  strategy,
-};
