--- packages/workbox-sw/_default.mjs
@@ -1,12 +1,0 @@
-/*
-  Copyright 2018 Google LLC
-
-  Use of this source code is governed by an MIT-style
-  license that can be found in the LICENSE file or at
-  https://opensource.org/licenses/MIT.
-*/
-
-import WorkboxSW from './controllers/WorkboxSW.mjs';
-import './_version.mjs';
-
-export default new WorkboxSW();
